requirements.txt
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
